requirements.txt
django~=3.2.18
django-cors-headers
celery~=5.1.2
python-decouple
djangorestframework~=3.14.0
datetime
psycopg2-binary
Pillow~=8.4.0
redis~=4.3.6
python-dateutil
pygments
babel